+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 # CHANGELOG
 
-## v0.1.1
+## v0.2.0
 
-Relaxing the constraints on the `Plymio.Codi` and `Harnais` dependencies.
+### Bug Fixes
 
-Corrected the API reference link in the README
+Fixed API Reference link in the README.
+
+### Changed Functions
+
+The export functions for both `Harnais.Error` and
+`Harnais.Error.Status` now take an *opts* as the optional second
+argument. The *opts*, if any, are used to update (`update/2`) the exception before
+performing the export.
+
+The keys in the `:export_config` used by `Harnais.Error`'s default export function
+are now normalised so that aliases can be used e.g. `:m` for
+`:message`, and also validated: unknown keys cause an error.
+
+### New Functions
+
+`Harnais.Error.export_exception/2` takes an `Exception` and optional
+*opts* and creates an export, returning `{:ok, export}` or `{:error, error}`.
 
 ## v0.1.0